+++ apps/web/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/responses/ResponseNote.tsx
@@ -1,11 +1,12 @@
 "use client";
 
+import { updateResponseNoteAction } from "@/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/responses/actions";
+import { useProfile } from "@/lib/profile";
 import { addResponseNote } from "@/lib/responseNotes/responsesNotes";
 import { timeSince } from "@formbricks/lib/time";
 import { TResponseNote } from "@formbricks/types/v1/responses";
 import { Button } from "@formbricks/ui";
-import { PlusIcon } from "@heroicons/react/24/outline";
-import { MinusIcon } from "@heroicons/react/24/solid";
+import { MinusIcon, PencilIcon, PlusIcon } from "@heroicons/react/24/solid";
 import clsx from "clsx";
 import { Maximize2Icon } from "lucide-react";
 import { useRouter } from "next/navigation";
@@ -30,8 +31,11 @@ export default function ResponseNotes({
   setIsOpen,
 }: ResponseNotesProps) {
   const router = useRouter();
+  const { profile } = useProfile();
   const [noteText, setNoteText] = useState("");
   const [isCreatingNote, setIsCreatingNote] = useState(false);
+  const [isUpdatingNote, setIsUpdatingNote] = useState(false);
+  const [noteId, setNoteId] = useState("");
   const divRef = useRef<HTMLDivElement>(null);
 
   const handleNoteSubmission = async (e: FormEvent) => {
@@ -45,6 +49,26 @@ export default function ResponseNotes({
     } catch (e) {
       toast.error("An error occurred creating a new note");
       setIsCreatingNote(false);
+    }
+  };
+
+  const handleEditPencil = (note: TResponseNote) => {
+    setNoteText(note.text);
+    setIsUpdatingNote(true);
+    setNoteId(note.id);
+  };
+
+  const handleNoteUpdate = async (e: FormEvent) => {
+    e.preventDefault();
+    setIsUpdatingNote(true);
+    try {
+      await updateResponseNoteAction(responseId, noteId, noteText);
+      router.refresh();
+      setIsUpdatingNote(false);
+      setNoteText("");
+    } catch (e) {
+      toast.error("An error occurred updating a note");
+      setIsUpdatingNote(false);
     }
   };
 
@@ -123,13 +147,20 @@ export default function ResponseNotes({
                     {timeSince(note.updatedAt.toISOString())}
                   </time>
                 </span>
-                <span className="block text-slate-700">{note.text}</span>
+                <div className="group/notetext flex items-center">
+                  <span className="block pr-1 text-slate-700">{note.text}</span>
+                  {profile.id === note.user.id && (
+                    <button onClick={() => handleEditPencil(note)}>
+                      <PencilIcon className=" h-3 w-3 text-gray-500 opacity-0 group-hover/notetext:opacity-100" />
+                    </button>
+                  )}
+                </div>
               </div>
             ))}
           </div>
           <div className="h-[120px]">
             <div className={clsx("absolute bottom-0 w-full px-3 pb-3", !notes.length && "absolute bottom-0")}>
-              <form onSubmit={handleNoteSubmission}>
+              <form onSubmit={isUpdatingNote ? handleNoteUpdate : handleNoteSubmission}>
                 <div className="mt-4">
                   <textarea
                     rows={2}
@@ -140,14 +171,16 @@ export default function ResponseNotes({
                     onKeyDown={(e) => {
                       if (e.key === "Enter" && noteText) {
                         e.preventDefault();
-                        handleNoteSubmission(e);
+                        {
+                          isUpdatingNote ? handleNoteUpdate(e) : handleNoteSubmission(e);
+                        }
                       }
                     }}
                     required></textarea>
                 </div>
                 <div className="mt-2 flex w-full justify-end">
                   <Button variant="darkCTA" size="sm" type="submit" loading={isCreatingNote}>
-                    Send
+                    {isUpdatingNote ? "Save" : "Send"}
                   </Button>
                 </div>
               </form>