+++ packages/js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @formbricks/js
 
+## 1.0.6
+
+### Patch Changes
+
+- 8efb1054: Introduce response queue for instant question transitions
+
+## 1.0.5
+
+### Patch Changes
+
+- bea1f993: Fix submit error in multiple choice questions
+
 ## 1.0.4
 
 ### Patch Changes